ML-354 Add imports to code generated by macros
The recommended fix of IMPORT ML_Core.Types AS %types% was not needed in this case.  The reference was changed from Types... to ML_Core.Types because the user had to import ML_Core to get the macro.  The other changes were to add imports to the Std.System.ThorLib module.

Signed-off-by: johnholt <john.d.holt@lexisnexis.com>